# `<algorithm>`: Optimize `sample()` and `shuffle()` with Lemire's algorithm