--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -6101,7 +6101,7 @@ private:
     _Udiff _Get_bits() { // return a random value within [0, _Bmask]
         constexpr auto _Urng_min = (_Urng::min) ();
         for (;;) { // repeat until random value is in range
-            const _Udiff _Val = _Ref() - _Urng_min;
+            const _Udiff _Val = static_cast<_Udiff>(_Ref() - _Urng_min);
 
             if (_Val <= _Bmask) {
                 return _Val;