--- a/stl/inc/random
+++ b/stl/inc/random
@@ -403,15 +403,6 @@ bool _Nrand_for_tr1(
     return false;
 }
 
-template <class _Gen, class = void>
-constexpr bool _Has_static_min_max = false;
-
-// This checks a requirement of N4981 [rand.req.urng] `concept uniform_random_bit_generator` but doesn't attempt
-// to implement the whole concept - we just need to distinguish Standard machinery from tr1 machinery.
-template <class _Gen>
-constexpr bool _Has_static_min_max<_Gen, void_t<decltype(bool_constant<(_Gen::min) () < (_Gen::max) ()>::value)>> =
-    true;
-
 template <class _Real, class _Gen>
 _NODISCARD _Real _Nrand_impl(_Gen& _Gx) { // build a floating-point value from random sequence
     _RNG_REQUIRE_REALTYPE(_Nrand_impl, _Real);